swahili/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai>=1.37.0,<2.0.0
pydantic-settings==2.7.1
websockets>=14.0
